Ensure footer is at the bottom of the pages
Sets the body to be full-screen height, with the background image
attached to it so it covers everything.

The body is also a flex container containing nav, content and footer in
a column. An empty spacer div with class flex-1 is also added between
the content div and the footer. This will expand to fill the available
space ensuring the footer is at the bottom of the page.