Add template for help videos
This removes a lot of repetition in the guidance.html page. The
video titles and URLs are now listed in the help_videos.txt file.
The build file runs the template against each entry in this file,
storing the output in an environment variable. The variable is
expanded by using the envsubst command on the content file in the
default template.