Redo home page story tiles to remove inline styles
These cause CSP to fail on a live site with "style-src 'self'" and we
don't want to add "unsafe-inline" if it isn't necessary.

There's no reason why a background image is needed for the tiles so
the style attributes have been removed and a plain img tag is used
instead.