Add separate CSS file for static site
This allows the CSS to be built separately from the app which is
currently important since we are using purgecss (and classes may
be accidentally removed which are used in the static site).

The background image is now added in the CSS file which gets
round the CSP banning of inline styles.