Hide horizontal overflow on body
When a vertical scrollbar is needed, the browser also adds a
horizontal one which isn't needed. This has something to do with
the extra width of the vertical scrollbar but I don't understand
exactly why it happens yet. In any case, adding

overflow-x: hidden;

for the body gets rid of it.